postfix services were added to Services class sin and cos functions were added
